move hover state in to uistate, pull out deployment hooks
